+++ Templates/src/Core/PainKiller.PowerCommands.Configuration/DOMAINOBJECTS/ConfigurationGlobals.cs
@@ -6,6 +6,7 @@ public static class ConfigurationGlobals
     public const string MainConfigurationFile = "PowerCommandsConfiguration.yaml";
     public const string SecurityFileName = "security.yaml";
     public const string WhatsNewFileName = "whats_new.md";
+    public const char ArraySplitter = '|';
 
     public static readonly string ApplicationDataFolder = $"{Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData)}\\{nameof(PowerCommands)}";
 }