+++ Templates/src/Core/PainKiller.PowerCommands.Configuration/EXTENSIONS/ConfigurationExtension.cs
@@ -8,7 +8,7 @@ public static class ConfigurationExtension
 
     public static string GetPath(this ArtifactPathsConfiguration configuration, string config)
     {
-        var paths = config.Split('|');
+        var paths = config.Split(ConfigurationGlobals.ArraySplitter);
         if (string.IsNullOrEmpty(config)) return "";
         if (paths.Length == 1) return ReplaceTags(config, configuration.Name);
         if (paths.Length == 2) return Path.Combine(ReplaceTags(paths[0], configuration.Name), ReplaceTags(paths[1], configuration.Name));