+++ Templates/src/Core/PainKiller.PowerCommands.ReadLine/EVENTS/CmdLineTextChangedArgs.cs
@@ -1,0 +1,6 @@
+﻿namespace $safeprojectname$.Events;
+public class CmdLineTextChangedArgs : EventArgs
+{
+    public CmdLineTextChangedArgs(string cmdText) => CmdText = cmdText;
+    public string CmdText { get; }
+}