+++ Templates/src/Core/PainKiller.PowerCommands.ReadLine/EVENTS/CommandHighlightedArgs.cs
@@ -1,0 +1,7 @@
+﻿namespace $safeprojectname$.Events;
+
+public class CommandHighlightedArgs : EventArgs
+{
+    public CommandHighlightedArgs(string commandName) => CommandName = commandName;
+    public string CommandName { get; }
+}