+++ Templates/src/PainKiller.PowerCommands.PowerCommandsConsole/Program.cs
@@ -1,4 +1,5 @@
-﻿using PainKiller.PowerCommands.Core.Services;
+﻿using System.Reflection;
+using PainKiller.PowerCommands.Core.Services;
 
 ConsoleService.Service.WriteLine(nameof(Program),@"______                        _____                                           _     
 | ___ \                      /  __ \                                         | |    
@@ -6,6 +7,5 @@ ConsoleService.Service.WriteLine(nameof(Program),@"______                       
 |  __/ _ \ \ /\ / / _ \ '__| | |    / _ \| '_ ` _ \| '_ ` _ \ / _` | '_ \ / _` / __|
 | | | (_) \ V  V /  __/ |    | \__/\ (_) | | | | | | | | | | | (_| | | | | (_| \__ \
 \_|  \___/ \_/\_/ \___|_|     \____/\___/|_| |_| |_|_| |_| |_|\__,_|_| |_|\__,_|___/", ConsoleColor.Cyan);
-ConsoleService.Service.WriteHeaderLine(nameof(Program),"\nVersion 1.0.0.1");
+ConsoleService.Service.WriteHeaderLine(nameof(Program),$"\nVersion {ReflectionService.Service.GetVersion(Assembly.GetExecutingAssembly())}");
 PainKiller.PowerCommands.Bootstrap.Startup.ConfigureServices().Run(args);
-