+++ src/PainKiller.PowerCommands/Core/PainKiller.PowerCommands.Core/PowerCommandsRuntime.cs
@@ -24,10 +24,10 @@ public class PowerCommandsRuntime<TConfig> : IPowerCommandsRuntime where TConfig
         {
             foreach (var command in proxyCommand.Commands)
             {
-                var identifiers = command.Split("|");
+                var identifiers = command.Split(ConfigurationGlobals.ArraySplitter);
                 var identifier = identifiers[0];
                 var identifierAlias = command.Contains($"|") ? identifiers[1] : identifier;
-                ConsoleService.Service.WriteLine("PowerCommandsRuntime", $"Proxy command [{identifierAlias}] added", null);
+                ConsoleService.Service.WriteLine("PowerCommandsRuntime", $"Proxy command [{identifierAlias}] added");
                 var powerCommand = new ProxyCommando(identifier, _configuration, proxyCommand.Name, proxyCommand.WorkingDirctory, identifierAlias);
                 SuggestionProviderManager.AddContextBoundSuggestions(identifierAlias, new[] { "--retry-interval-seconds", "--no-quit","--help" });
                 if(Commands.All(c => c.Identifier != powerCommand.Identifier)) Commands.Add(powerCommand);
@@ -54,14 +54,17 @@ public class PowerCommandsRuntime<TConfig> : IPowerCommandsRuntime where TConfig
             if (!attrib.OverrideHelpOption)
             {
                 HelpService.Service.ShowHelp(command, clearConsole: true);
+                command.RunCompleted();
                 return new RunResult(command, input, "User prompted for help with --help option", RunResultStatus.Ok);
             }
         }
         if (command.InitializeAndValidateInput(input, attrib))
         {
             Latest = new RunResult(command, input, "Validation error", RunResultStatus.InputValidationError);
+            command.RunCompleted();
             return Latest;
         }
+
         if (command.GetPowerCommandAttribute().UseAsync) return ExecuteAsyncCommand(command, input);
         try
         {
@@ -76,14 +79,17 @@ public class PowerCommandsRuntime<TConfig> : IPowerCommandsRuntime where TConfig
     {
         try
         {
-            command.RunAsync().ConfigureAwait(continueOnCapturedContext:false);
+            command.RunAsync().ContinueWith((_) =>
+            {
+                command.RunCompleted();
+                Console.Write(ConfigurationGlobals.Prompt);
+            });
+            Latest = new RunResult(command, input, "Command running async operation", RunResultStatus.Async);
         }
         catch (Exception e)
         {
             Latest = new RunResult(command, input, e.Message, RunResultStatus.ExceptionThrown);
-            return Latest;
         }
-        Latest = new RunResult(command, input, "Command running async operation", RunResultStatus.Async);
         return Latest;
     }
     public RunResult? Latest { get; private set; }